requirements.txt
pydantic>=2.10
pydantic-settings>=2.0.0
python-dotenv>=1.0
fastapi>=0.115
uvicorn>=0.34
llama-index-core>=0.13
llama-index-embeddings-openai>=0.5
llama-index-llms-openai>=0.4
llama-index-vector-stores-qdrant>=0.8
qdrant-client>=1.7.0,<1.12.0
pypdf>=5.1
openai>=1.50

# Development dependencies
ruff>=0.1.0
black>=23.0.0
mypy>=1.0.0